# Conflicts:
#	features/communication/topchat/src/main/java/com/tokopedia/topchat/chatroom/view/viewmodel/TopChatViewModel.kt

+++ features/transaction/cart/src/main/java/com/tokopedia/cart/view/CartListPresenter.kt
@@ -703,25 +703,25 @@ class CartListPresenter @Inject constructor(
 
         for (wholesalePriceData in wholesalePriceDataList) {
             if (itemQty >= wholesalePriceData.qtyMin) {
-                subTotalWholesalePrice = (itemQty * wholesalePriceData.prdPrc).toDouble()
+                subTotalWholesalePrice = (itemQty * wholesalePriceData.prdPrc)
                 hasCalculateWholesalePrice = true
                 val wholesalePriceFormatted = CurrencyFormatUtil.convertPriceValueToIdrFormat(
                     wholesalePriceData.prdPrc, false
                 ).removeDecimalSuffix()
                 cartItemHolderData.wholesalePriceFormatted = wholesalePriceFormatted
                 cartItemHolderData.wholesalePrice = wholesalePriceData.prdPrc
-                subtotalBeforeSlashedPrice = itemQty * cartItemHolderData.wholesalePrice.toDouble()
+                subtotalBeforeSlashedPrice = itemQty * cartItemHolderData.wholesalePrice
                 break
             }
         }
 
         if (!hasCalculateWholesalePrice) {
-            subTotalWholesalePrice = (itemQty * cartItemHolderData.productPrice).toDouble()
+            subTotalWholesalePrice = (itemQty * cartItemHolderData.productPrice)
             cartItemHolderData.wholesalePriceFormatted = null
-            cartItemHolderData.wholesalePrice = 0
+            cartItemHolderData.wholesalePrice = 0.0
             subtotalBeforeSlashedPrice =
-                if (cartItemHolderData.productOriginalPrice > 0) (itemQty * cartItemHolderData.productOriginalPrice).toDouble()
-                else (itemQty * cartItemHolderData.productPrice).toDouble()
+                if (cartItemHolderData.productOriginalPrice > 0) (itemQty * cartItemHolderData.productOriginalPrice)
+                else (itemQty * cartItemHolderData.productPrice)
         }
 
         if (cartItemHolderData.productCashBack.isNotBlank()) {
